requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.23

# Security
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# LINE Integration
line-bot-sdk==3.6.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0
